Gets scaffold to pass along relations, refactor relation lookup
